Rename bootstrap to polyfills.
